+++ samples/samples-aspnet/embedded-auth-with-sdk/embedded-auth-with-sdk/Models/OktaVerifyPollResponseModel.cs
@@ -1,0 +1,11 @@
+﻿namespace embedded_auth_with_sdk.Models
+{
+    public class OktaVerifyPollResponseModel
+    {
+        public int? Refresh { get; set; } 
+
+        public bool? ContinuePolling { get; set; }
+
+        public string Next { get; set; }
+    }
+}